requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
dataclasses==0.5
Django==2.0.4
django-bootstrap3==9.1.0
djangorestframework==3.8.1
djongo==1.2.23
idna==2.6
Keras==2.1.5
kiwisolver==1.0.1
matplotlib==2.2.2
mongoengine==0.8.0
numpy==1.14.2
pymongo==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
scipy==1.0.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
